Refactor of dynamodbDocClientInit for tidyness
